방문 인증 정책 변경, util 코드 분리 - 방문 인증, 점포 등록 유효 거리 100m로 조정 - Service 계층에서 util 클래스로 일부 로직 분리 - 연령/나이대 별 추천 점포 응답 타입 변경 - 테스트 케이스 작성
