+++ src/main/java/com/dongyang/dongpo/dto/store/StoreVisitCertDto.java
@@ -1,0 +1,13 @@
+package com.dongyang.dongpo.dto.store;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class StoreVisitCertDto {
+    private double latitude; // 위도
+    private double longitude; // 경도
+    private Long storeId;
+    private boolean isVisitSuccessful;
+}
